move over my face in svg, and add the hover effects
