+++ src/Face/Face.js
@@ -1,18 +1,20 @@
 import React from 'react';
 
+import styles from './Face.css';
+
 const Face = () =>
    (
-     <svg className="logo" id="logo seanFace" xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" viewbox="0 0 500 500" xmlSpace="preserve" aria-labelledby="title desc">
+     <svg className={styles.face} xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" viewbox="0 0 500 500" xmlSpace="preserve" aria-labelledby="title desc">
        <title>Sean Kilgarriff Face Logo</title>
        <desc id="desc">A face with hair, glasses, and a beard to represent Sean Kilgarriff</desc>
        <g>
 
-         <path className="beard" id="beard" fill="#323232" roll="button" d="M365.6,461.9c0-11.2,5.6-37.7,13.6-60.1c8-22.4-5.6-41.2-9.6-62.9c-4-21.7,6-51.7,0-55.2 c-9.6-5.6-11.2,11.9-15.1,33.5c-4,21.7-28.7,31.5-28.7,31.5s-12-13.3-25.5-21c-13.5-7.7-50.2-8.4-50.2-8.4s-36.7,0.7-50.2,8.4 c-13.5,7.7-25.5,21-25.5,21s-24.7-9.8-28.7-31.5c-4-21.7-5.6-39.1-15.1-33.5c-6,3.5,4,33.5,0,55.2c-4,21.7-17.5,40.5-9.6,62.9 c8,22.4,13.5,48.9,13.5,60.1c19.1,10.6,62.7,31.4,121.5,29.9C308,490.5,346.7,472.4,365.6,461.9z M250,529 M298.6,369 c-11.2,14-48.6,11.9-48.6,11.9s-37.5,2.1-48.6-11.9c-11.2-14,42.2,0.7,48.6-16.1C256.4,369.7,309.8,355,298.6,369z" />
+         <path className={styles.beard} fill="#323232" roll="button" d="M365.6,461.9c0-11.2,5.6-37.7,13.6-60.1c8-22.4-5.6-41.2-9.6-62.9c-4-21.7,6-51.7,0-55.2 c-9.6-5.6-11.2,11.9-15.1,33.5c-4,21.7-28.7,31.5-28.7,31.5s-12-13.3-25.5-21c-13.5-7.7-50.2-8.4-50.2-8.4s-36.7,0.7-50.2,8.4 c-13.5,7.7-25.5,21-25.5,21s-24.7-9.8-28.7-31.5c-4-21.7-5.6-39.1-15.1-33.5c-6,3.5,4,33.5,0,55.2c-4,21.7-17.5,40.5-9.6,62.9 c8,22.4,13.5,48.9,13.5,60.1c19.1,10.6,62.7,31.4,121.5,29.9C308,490.5,346.7,472.4,365.6,461.9z M250,529 M298.6,369 c-11.2,14-48.6,11.9-48.6,11.9s-37.5,2.1-48.6-11.9c-11.2-14,42.2,0.7,48.6-16.1C256.4,369.7,309.8,355,298.6,369z" />
 
 
-         <path pointerEvents="bounding-box" className="glasses" id="glasses" fill="#323232" d="M337.1,212h-66.4c-6.5,0-11.8,5.3-11.8,11.8v4.7c-3.2-2.8-6.2-4.3-8.9-4.3c-2.7,0-5.7,1.4-8.9,4.3v-4.7 c0-6.5-5.3-11.8-11.8-11.8h-66.4c-6.5,0-11.8,5.3-11.8,11.8v27c0,6.5,5.3,11.8,11.8,11.8h66.4c6.5,0,11.8-5.3,11.8-11.8v-14.1 c2.7-3.3,6.6-6.8,8.9-6.8c2.2,0,6.2,3.4,8.9,6.8v14.1c0,6.5,5.3,11.8,11.8,11.8h66.4c6.5,0,11.8-5.3,11.8-11.8v-27 C349,217.3,343.7,212,337.1,212z M235.4,250.8c0,3.4-2.8,6.1-6.1,6.1h-66.4c-3.4,0-6.1-2.8-6.1-6.1v-27c0-3.4,2.8-6.1,6.1-6.1 h66.4c3.4,0,6.1,2.8,6.1,6.1L235.4,250.8L235.4,250.8z M343.3,250.8c0,3.4-2.8,6.1-6.1,6.1h-66.4c-3.4,0-6.1-2.8-6.1-6.1v-27 c0-3.4,2.8-6.1,6.1-6.1h66.4c3.4,0,6.1,2.8,6.1,6.1V250.8z" />
+         <path pointerEvents="bounding-box" className={styles.glasses} fill="#323232" d="M337.1,212h-66.4c-6.5,0-11.8,5.3-11.8,11.8v4.7c-3.2-2.8-6.2-4.3-8.9-4.3c-2.7,0-5.7,1.4-8.9,4.3v-4.7 c0-6.5-5.3-11.8-11.8-11.8h-66.4c-6.5,0-11.8,5.3-11.8,11.8v27c0,6.5,5.3,11.8,11.8,11.8h66.4c6.5,0,11.8-5.3,11.8-11.8v-14.1 c2.7-3.3,6.6-6.8,8.9-6.8c2.2,0,6.2,3.4,8.9,6.8v14.1c0,6.5,5.3,11.8,11.8,11.8h66.4c6.5,0,11.8-5.3,11.8-11.8v-27 C349,217.3,343.7,212,337.1,212z M235.4,250.8c0,3.4-2.8,6.1-6.1,6.1h-66.4c-3.4,0-6.1-2.8-6.1-6.1v-27c0-3.4,2.8-6.1,6.1-6.1 h66.4c3.4,0,6.1,2.8,6.1,6.1L235.4,250.8L235.4,250.8z M343.3,250.8c0,3.4-2.8,6.1-6.1,6.1h-66.4c-3.4,0-6.1-2.8-6.1-6.1v-27 c0-3.4,2.8-6.1,6.1-6.1h66.4c3.4,0,6.1,2.8,6.1,6.1V250.8z" />
 
-         <path className="hair" id="hair" fill="#323232" d="M142.7,150.2c-1.4,8.4-2.5,117.9-8.5,126.9c-6.1,9-40.5-104.4-34.5-159.7c1-9-14.3,3.1-17.3,16.1 c0,0-1.2-35,27.3-43.1c0,0-1.2-41.9,32.3-64.1c0,0-2,14-1.9,22c0,0,43.6-45.2,144.6-38.7c101,6.5,144.3,95.2,144.3,95.2 S417.8,93.9,409.6,93c0,0-24.5,159-42.9,179c-10.1,11,4.7-83.9-7.7-105.8C346.6,144.2,203.7,112,190.4,115 C177.2,118.1,144.6,138.2,142.7,150.2z" />
+         <path className={styles.hair} fill="#323232" d="M142.7,150.2c-1.4,8.4-2.5,117.9-8.5,126.9c-6.1,9-40.5-104.4-34.5-159.7c1-9-14.3,3.1-17.3,16.1 c0,0-1.2-35,27.3-43.1c0,0-1.2-41.9,32.3-64.1c0,0-2,14-1.9,22c0,0,43.6-45.2,144.6-38.7c101,6.5,144.3,95.2,144.3,95.2 S417.8,93.9,409.6,93c0,0-24.5,159-42.9,179c-10.1,11,4.7-83.9-7.7-105.8C346.6,144.2,203.7,112,190.4,115 C177.2,118.1,144.6,138.2,142.7,150.2z" />
        </g>
      </svg>
   )
