+++ src/Home/Home.css
@@ -1,0 +1,4 @@
+.container {
+  height: 100%;
+  width: 100%;
+}
