+++ src/Home/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "Home",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./Home.js"
+}
