--- src/assets/styles/_footer.scss
@@ -1,34 +1,0 @@
-// -----------------------------------------------------------------------------
-// This file contains all styles related to the footer of the site/application.
-// -----------------------------------------------------------------------------
-
-$link-base-color: #323232;
-
-footer {
-  .copyright,
-  .social {
-    @include span-columns(12);
-    font-size: 0.8em;
-    text-align: center;
-
-    @include media($medium-screen) {
-      font-size: 1em;
-    }
-  }
-
-  .copyright {
-    @include media($medium-screen) {
-      margin-top: 10px;
-    }
-
-    a {
-      color: $base-font-color;
-      display: inline-block;
-    }
-
-    #PGP {
-      margin-left: 20px;
-    }
-
-  }
-}
