--- src/partials/footer.hbs
@@ -1,59 +1,0 @@
-<footer class="footer">
-<!-- TODO: Order the Social Media Icons how I want Mail middle.  -->
-				<div class="social">
-						<a href="https://medium.com/@SKilgarriff" target="_blank">
-								<span class="fa-stack fa-lg">
-										<i class="fa fa-circle fa-stack-2x"></i>
-										<i class="fa fa-medium fa-stack-1x fa-inverse"></i>
-								</span>
-						</a>
-						<a href="https://www.facebook.com/skilgarriff" target="_blank">
-								<span class="fa-stack fa-lg">
-										<i class="fa fa-circle fa-stack-2x"></i>
-										<i class="fa fa-facebook fa-stack-1x fa-inverse"></i>
-								</span>
-						</a>
-						<a href="https://twitter.com/SeanKilgarriff" target="_blank">
-								<span class="fa-stack fa-lg">
-										<i class="fa fa-circle fa-stack-2x"></i>
-										<i class="fa fa-twitter fa-stack-1x fa-inverse"></i>
-								</span>
-						</a>
-						<a href="mailto:sean@zzrot.com">
-								<span class="fa-stack fa-lg">
-										<i class="fa fa-circle fa-stack-2x"></i>
-										<i class="fa fa-envelope fa-stack-1x fa-inverse"></i>
-								</span>
-						</a>
-						<a href="https://github.com/Skilgarriff" target="_blank">
-								<span class="fa-stack fa-lg">
-										<i class="fa fa-circle fa-stack-2x"></i>
-										<i class="fa fa-github-alt fa-stack-1x fa-inverse"></i>
-								</span>
-						</a>
-						<a href="https://www.linkedin.com/in/seankilgarriff" target="_blank">
-								<span class="fa-stack fa-lg">
-										<i class="fa fa-circle fa-stack-2x"></i>
-										<i class="fa fa-linkedin fa-stack-1x fa-inverse"></i>
-								</span>
-						</a>
-						<a href="https://www.instagram.com/skilgarriff/" target="_blank">
-								<span class="fa-stack fa-lg">
-										<i class="fa fa-circle fa-stack-2x"></i>
-										<i class="fa fa-instagram fa-stack-1x fa-inverse"></i>
-								</span>
-						</a>
-				</div>
-				<div class="copyright">
-						<a href="https://seankilgarriff.com/copyright">
-								<p>
-										&copy; Sean Kilgarriff 2016
-								</p>
-						</a>
-						<a href="https://seankilgarriff.com/pgp" id="PGP">
-                            <p>
-                                PGP Key
-                            </p>
-                        </a>
-				</div>
-		</footer>
